requirements.txt
matplotlib==3.7.1
numpy==1.23.5
openai==0.27.6
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.22.0
